requirement.txt
django==2.2
djangorestframework==3.9.2
djongo
sqlparse==0.2.4